getting an early benchmark suite up that compare raw UnixFS, WNFS public, and WNFS private. Results so far:

```
goos: darwin
goarch: amd64
pkg: github.com/qri-io/wnfs-go
cpu: Intel(R) Core(TM) i7-6567U CPU @ 3.30GHz

BenchmarkIPFSCat10MbFile-4                             	   61572	     20811 ns/op	   49822 B/op	      42 allocs/op
BenchmarkIPFSWrite10MbFile-4                           	   30804	     47417 ns/op	   11432 B/op	     204 allocs/op
BenchmarkIPFSCat10MbFileSubdir-4                       	   36136	     30758 ns/op	   53418 B/op	     123 allocs/op
BenchmarkIPFSWrite10MbFileSubdir-4                     	   14757	     85009 ns/op	   27069 B/op	     524 allocs/op
BenchmarkIPFSCp10DirectoriesWithOne10MbFileEach-4      	     652	   1832513 ns/op	  253688 B/op	    2472 allocs/op

BenchmarkPublicCat10MbFile-4                           	   34389	     34300 ns/op	   53607 B/op	     108 allocs/op
BenchmarkPublicWrite10MbFile-4                         	    4340	    251079 ns/op	   66669 B/op	    1168 allocs/op
BenchmarkPublicCat10MbFileSubdir-4                     	   20268	     52665 ns/op	   62005 B/op	     243 allocs/op
BenchmarkPublicWrite10MbFileSubdir-4                   	    2546	    417772 ns/op	  116442 B/op	    2029 allocs/op
BenchmarkPublicCp10DirectoriesWithOne10MbFileEach-4    	  238276	      5297 ns/op	    1156 B/op	      19 allocs/op

BenchmarkPrivateCat10MbFile-4                          	   23559	     61366 ns/op	  115821 B/op	     119 allocs/op
BenchmarkPrivateWrite10MbFile-4                        	    2440	    515472 ns/op	  156549 B/op	    2287 allocs/op
BenchmarkPrivateCat10MbFileSubdir-4                    	   17660	     68230 ns/op	  124090 B/op	     187 allocs/op
BenchmarkPrivateWrite10MbFileSubdir-4                  	    1707	    755976 ns/op	  228413 B/op	    3289 allocs/op
BenchmarkPrivateCp10DirectoriesWithOne10MbFileEach-4   	     162	   6474506 ns/op	 1246602 B/op	   10928 allocs/op
```